+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.10
 //
 //  Copyright 2021 Readium Foundation. All rights reserved.
 //  Use of this source code is governed by the BSD-style license
@@ -24,23 +24,23 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/krzyzanowskim/CryptoSwift.git", from: "1.8.0"),
-        .package(url: "https://github.com/marmelroy/Zip.git", from: "2.1.0"),
         .package(url: "https://github.com/ra1028/DifferenceKit.git", from: "1.3.0"),
-        .package(url: "https://github.com/readium/Fuzi.git", from: "3.1.4"),
+        .package(url: "https://github.com/readium/Fuzi.git", from: "4.0.0"),
         .package(url: "https://github.com/readium/GCDWebServer.git", from: "4.0.0"),
+        .package(url: "https://github.com/readium/ZIPFoundation.git", from: "1.0.0"),
         .package(url: "https://github.com/scinfu/SwiftSoup.git", from: "2.7.0"),
         .package(url: "https://github.com/stephencelis/SQLite.swift.git", from: "0.15.0"),
-        .package(url: "https://github.com/weichsel/ZIPFoundation.git", from: "0.9.0"),
     ],
     targets: [
         .target(
             name: "ReadiumShared",
-            dependencies: ["ReadiumInternal", "Fuzi", "SwiftSoup", "Zip"],
-            path: "Sources/Shared",
-            exclude: [
-                // Support for ZIPFoundation is not yet achieved.
-                "Toolkit/ZIP/ZIPFoundation.swift",
+            dependencies: [
+                "ReadiumInternal",
+                "SwiftSoup",
+                .product(name: "ReadiumFuzi", package: "Fuzi"),
+                .product(name: "ReadiumZIPFoundation", package: "ZIPFoundation"),
             ],
+            path: "Sources/Shared",
             linkerSettings: [
                 .linkedFramework("CoreServices"),
                 .linkedFramework("UIKit"),
@@ -59,8 +59,8 @@ let package = Package(
             name: "ReadiumStreamer",
             dependencies: [
                 "CryptoSwift",
-                "Fuzi",
                 "ReadiumShared",
+                .product(name: "ReadiumFuzi", package: "Fuzi"),
             ],
             path: "Sources/Streamer",
             resources: [
@@ -102,8 +102,8 @@ let package = Package(
         .target(
             name: "ReadiumOPDS",
             dependencies: [
-                "Fuzi",
                 "ReadiumShared",
+                .product(name: "ReadiumFuzi", package: "Fuzi"),
             ],
             path: "Sources/OPDS"
         ),
@@ -120,8 +120,8 @@ let package = Package(
             name: "ReadiumLCP",
             dependencies: [
                 "CryptoSwift",
-                "ZIPFoundation",
                 "ReadiumShared",
+                .product(name: "ReadiumZIPFoundation", package: "ZIPFoundation"),
             ],
             path: "Sources/LCP",
             resources: [